refactor: validation and review (module 3) (#89)
* refactor: decompose submission-detail and manual-entry, extract shared styles and hooks

* fix: stable useCallback deps, remove unused showRR, memoize handleRemoveImage

---------
