+++ src/features/validation/components/validation-input-styles.ts
@@ -1,0 +1,12 @@
+import { mflColors } from '../../../constants/colors';
+
+export const pointsInputStyle = {
+  backgroundColor: mflColors.card,
+  borderWidth: 1,
+  borderColor: mflColors.border,
+  borderRadius: 10,
+  paddingHorizontal: 12,
+  paddingVertical: 9,
+  fontSize: 14,
+  color: mflColors.text,
+};
